requirements.txt
discord.py==2.5.2
python-dotenv>=1.0.0
aiohttp==3.9.1
asyncio==3.4.3
typing-extensions==4.8.0
pydantic==2.5.2
asyncpg>=0.29.0 

# LLM Providers
openai==1.3.7
anthropic==0.8.1
google-generativeai==0.3.1

# Add tenacity for retry logic
tenacity>=8.0.0

aiomysql>=0.0.21
PyYAML>=6.0 